+++ dynamic_instrumentation/src/main.rs
@@ -2,6 +2,18 @@
 extern crate rustc_driver;
 extern crate rustc_session;
 extern crate rustc_span;
+extern crate rustc_ast;
+extern crate rustc_const_eval;
+extern crate rustc_data_structures;
+extern crate rustc_index;
+extern crate rustc_interface;
+extern crate rustc_middle;
+extern crate rustc_mir_build;
+extern crate rustc_mir_transform;
+extern crate rustc_target;
+
+mod instrument_memory;
+use instrument_memory::InstrumentMemoryOps;
 
 use std::{
     env,
@@ -10,12 +22,27 @@ use std::{
     process::{self, Command, ExitStatus},
 };
 
+use rustc_ast::ast::{Item, ItemKind, Visibility, VisibilityKind};
+use rustc_ast::node_id::NodeId;
+use rustc_ast::ptr::P;
+use rustc_const_eval::transform::validate;
+use rustc_driver::Compilation;
+use rustc_driver::{RunCompiler, TimePassesCallbacks};
+use rustc_interface::interface::Compiler;
+use rustc_interface::Queries;
+use rustc_middle::mir::MirPass;
+use rustc_middle::ty::query::{ExternProviders, Providers};
+use rustc_middle::ty::WithOptConstParam;
+use rustc_session::Session;
+use rustc_span::def_id::LocalDefId;
+use rustc_span::symbol::Ident;
+use rustc_span::DUMMY_SP;
+
 use anyhow::anyhow;
-use c2rust_dynamic_instrumentation::{MirTransformCallbacks, INSTRUMENTER};
 use camino::Utf8Path;
 use cargo_metadata::MetadataCommand;
 use clap::Parser;
-use rustc_driver::{RunCompiler, TimePassesCallbacks};
+use lazy_static::lazy_static;
 
 /// Instrument memory accesses for dynamic analysis.
 #[derive(Debug, Parser)]
@@ -49,7 +76,7 @@ fn get_sysroot_slow() -> anyhow::Result<PathBuf> {
 }
 
 /// Resolve the current `rustc` sysroot.
-/// 
+///
 /// Normally, `rustc` looks up the sysroot by the location of its own binary.
 /// This works because the `rustc` on `$PATH` is actually `rustup`,
 /// and `rustup` invokes the real `rustc`, which is in a location relative to the sysroot.
@@ -134,4 +161,67 @@ fn main() -> anyhow::Result<()> {
         exit_with_status(status);
     }
     Ok(())
+}
+
+lazy_static! {
+    static ref INSTRUMENTER: InstrumentMemoryOps = InstrumentMemoryOps::new();
+}
+
+struct MirTransformCallbacks;
+
+impl rustc_driver::Callbacks for MirTransformCallbacks {
+    fn config(&mut self, config: &mut rustc_interface::Config) {
+        config.override_queries = Some(override_queries);
+    }
+
+    fn after_parsing<'tcx>(
+        &mut self,
+        _compiler: &Compiler,
+        queries: &'tcx Queries<'tcx>,
+    ) -> Compilation {
+        let parse = queries.parse().unwrap();
+        let mut parse = parse.peek_mut();
+        parse.items.push(P(Item {
+            attrs: Vec::new(),
+            id: NodeId::from_u32(0),
+            span: DUMMY_SP,
+            vis: Visibility {
+                kind: VisibilityKind::Inherited,
+                span: DUMMY_SP,
+                tokens: None,
+            },
+            ident: Ident::from_str("c2rust_analysis_rt"),
+            kind: ItemKind::ExternCrate(None),
+            tokens: None,
+        }));
+        Compilation::Continue
+    }
+}
+
+fn override_queries(
+    _sess: &Session,
+    providers: &mut Providers,
+    _extern_providers: &mut ExternProviders,
+) {
+    providers.mir_built = |tcx, def: WithOptConstParam<LocalDefId>| {
+        let mut providers = Providers::default();
+        rustc_mir_build::provide(&mut providers);
+
+        let steal_mir = (providers.mir_built)(tcx, def);
+        let mut mir = steal_mir.steal();
+
+        let body_did = def.did.to_def_id();
+        let fn_ty = tcx.type_of(body_did);
+        if fn_ty.is_fn() && !tcx.is_const_fn(body_did) && !tcx.is_static(body_did) {
+            INSTRUMENTER.instrument_fn(tcx, &mut mir, body_did);
+
+            validate::Validator {
+                when: "After dynamic instrumentation".to_string(),
+                mir_phase: mir.phase,
+            }
+            .run_pass(tcx, &mut mir);
+        }
+
+        tcx.alloc_steal_mir(mir)
+    };
 }
